+++ macOS/PrivateFrameworks/SiriUI.framework/SiriUIConfiguration.h
@@ -11,13 +11,13 @@
 @interface SiriUIConfiguration : NSObject <NSCopying>
 {
     BOOL _isSystemHostedPresentation;
-    int _flamesViewFidelity;
+    long long _flamesViewFidelity;
 }
 
 @property(readonly, nonatomic) BOOL isSystemHostedPresentation; // @synthesize isSystemHostedPresentation=_isSystemHostedPresentation;
-@property(readonly, nonatomic) int flamesViewFidelity; // @synthesize flamesViewFidelity=_flamesViewFidelity;
+@property(readonly, nonatomic) long long flamesViewFidelity; // @synthesize flamesViewFidelity=_flamesViewFidelity;
 - (id)copyWithZone:(struct _NSZone *)arg1;
-- (id)initWithFlamesViewFidelity:(int)arg1 isSystemHostedPresentation:(BOOL)arg2;
+- (id)initWithFlamesViewFidelity:(long long)arg1 isSystemHostedPresentation:(BOOL)arg2;
 - (id)init;
 
 @end
